POSTAGGER with Handcraft Feature(50%) 1. 因为一些明显的错误耽误大量的时间——模板函数、类定定义必须在头文件中，否则报找不到定义。当编译器一直提示链接出错，找不到定义时，我始终没有发现问题。 2. 写的过程中大量错误——定义hpp文件却却不将函数定义为inline。hpp是有局限的，有static变量初始化的话就必须定义到头文件中，同时为了写HPP而将不该定义为inline的定义为inline，非常不合理。 3. 大量拼写错误。需要考虑将名字取得短一些？

+++ src/postagger/postagger_module/pos_feature_layer.cpp
@@ -3,10 +3,10 @@
 namespace slnn{
 
 POSFeatureLayer::POSFeatureLayer(cnn::Model *m, 
-                                 size_t prefix_suffix_len1_dict_size, size_t prefix_suffix_len1_embedding_dim,
-                                 size_t prefix_suffix_len2_dict_size, size_t prefix_suffix_len2_embedding_dim,
-                                 size_t prefix_suffix_len3_dict_size, size_t prefix_suffix_len3_embedding_dim,
-                                 size_t char_length_dict_size, size_t char_length_embedding_dim)
+                                 size_t prefix_suffix_len1_dict_size, unsigned prefix_suffix_len1_embedding_dim,
+                                 size_t prefix_suffix_len2_dict_size, unsigned prefix_suffix_len2_embedding_dim,
+                                 size_t prefix_suffix_len3_dict_size, unsigned prefix_suffix_len3_embedding_dim,
+                                 size_t char_length_dict_size, unsigned char_length_embedding_dim)
     :prefix_suffix_len1_lookup_param(m->add_lookup_parameters(prefix_suffix_len1_dict_size, {prefix_suffix_len1_embedding_dim})),
     prefix_suffix_len2_lookup_param(m->add_lookup_parameters(prefix_suffix_len2_dict_size, {prefix_suffix_len2_embedding_dim})),
     prefix_suffix_len3_lookup_param(m->add_lookup_parameters(prefix_suffix_len3_dict_size, {prefix_suffix_len3_embedding_dim})),
